split GUI and tag management out of Element into separate classes
